requirements.txt
requests==2.32.3
PyYAML==6.0.1
google-generativeai==0.5.4
Pillow==10.4.0
aiofiles==24.1.0
flask==3.0.0
paho-mqtt==2.1.0
psutil==5.9.6